+++ packages/eslint-config-d3banking/CHANGELOG.md
@@ -2,6 +2,8 @@
 ==================
 - [deps] Update `eslint` to `v3.2`
 - [breaking] Update `eslint-config-airbnb-base` to `v5.0.1`
+- [minor] Add d3 specific overrides for new rules
+- [breaking] Remove `angular`, `d3`, and `$` as globals
 
 7.1.3 / 2016-05-26
 ==================